+++ timemachine/cpp/src/harmonic_bond.cu
@@ -73,14 +73,22 @@ void HarmonicBond<RealType>::execute_device(
     const double lambda,
     unsigned long long *d_du_dx,
     double *d_du_dp,
-    double *d_du_dl,
-    double *d_u,
+    unsigned long long *d_du_dl,
+    unsigned long long *d_u,
     cudaStream_t stream) {
 
     int tpb = 32;
     int blocks = (B_+tpb-1)/tpb;
 
-    if(B_ > 0) { 
+    // if(d_du_dl) {
+    //     this->reset_du_dl_buffer(stream);
+    // }
+
+    // if(d_u) {
+    //     this->reset_u_buffer(stream);
+    // }
+
+    if(B_ > 0) {
         k_harmonic_bond<RealType><<<blocks, tpb, 0, stream>>>(
             B_,
             d_x,
@@ -96,6 +104,15 @@ void HarmonicBond<RealType>::execute_device(
         );
         gpuErrchk(cudaPeekAtLastError());
     }
+
+    // if(d_du_dl) {
+    //     this->reduce_du_dl_buffer(d_du_dl, stream);
+    // }
+
+    // if(d_u) {
+    //     this->reduce_u_buffer(d_u, stream);
+    // }
+
 };
 
 template class HarmonicBond<double>;
